Changes in ParamGen

+++ Cryptosystem/src/decrypt/CiphertextDecrypter.java
@@ -1,17 +1,8 @@
 package decrypt;
 
-import java.math.BigInteger;
-
-/**
- * 
- * @author Aaron
- * Contains method to decrypt ciphertext and related helper-methods.
- *
- */
-
 public class CiphertextDecrypter {
 	
-	public BigInteger[][] decrypt(BigInteger[][] ciphertext, BigInteger[][] privateKey){
+	public long[][] decrypt(long[][] ciphertext, long[][] privateKey){
 		
 		return null;
