+++ Cryptosystem/src/encrypt/Encrypter.java
@@ -1,9 +1,7 @@
 package encrypt;
 
-import java.math.BigInteger;
-
 public interface Encrypter {
 
-	public BigInteger[][] encrypt(BigInteger[][] plaintext ,BigInteger[][] publicKey);
+	public long[][] encrypt(long[][] plaintext ,long[][] publicKey);
 	
 }
