+++ Cryptosystem/src/homomorphicOperations/HomomorphicOperations.java
@@ -1,8 +1,6 @@
 package homomorphicOperations;
 
-import java.math.BigInteger;
-
 public interface HomomorphicOperations {
-	public BigInteger[][] add(BigInteger[][] c1, BigInteger[][] c2);
-	public BigInteger[][] multiply(BigInteger[][] c1,BigInteger[][] c2);
+	public long[][] add(long[][] c1, long[][] c2);
+	public long[][] multiply(long[][] c1,long[][] c2);
 }
